Move RootBuildCacheControllerRef to :core
It's a workaround for service scopes, nothing to do with the build-cache controller per se.
